Feat : 스크럼 서비스 추상화, 스크럼 참여자 QueryDsl로 변경
