+++ backend/api/src/main/java/com/parkchoi/scrum/domain/scrum/repository/scrumparticipant/ScrumParticipantRepository.java
@@ -1,0 +1,10 @@
+package com.parkchoi.scrum.domain.scrum.repository.scrumparticipant;
+
+import com.parkchoi.scrum.domain.scrum.entity.ScrumParticipant;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface ScrumParticipantRepository extends JpaRepository<ScrumParticipant, Long>, ScrumParticipantCustom {
+}
